.editorconfig update
Generated via Intellij IDEA and so has a bunch of Intellij IDEA specific options. Not sure what to do about this as the base editorconfig options are not powerful enough to handle a standardised style as comprehensive as this, yet it probably wont work too well with Eclipse. Will have to see, worst case scenario is just reverting this an exporting these options as an IDEA config file, but I figured I'd try this first.